+++ fastquant/__init__.py
@@ -1,2 +1,3 @@
 from .fastquant import *
+from .disclosures import *
 from .strategies import *
