+++ fastquant/fastquant.py
@@ -13,7 +13,6 @@ import pandas as pd
 import numpy as np
 from string import digits
 import lxml.html as LH
-from bs4 import BeautifulSoup
 from tqdm import tqdm
 import tweepy
 from pathlib import Path
@@ -493,220 +492,6 @@ def pse_data_to_csv(
         performance_dict["E"].to_csv(
             "{}{}_{}_{}_E.csv".format(pse_dir, symbol, start_date, end_date)
         )
-
-
-def get_company_disclosures(
-    symbol, from_date="06-26-2017", to_date="06-26-2019"
-):
-    """
-    symbol str - Ticker of the pse stock of choice
-    from_date date str %m-%d-%Y - Beginning date of the disclosure data pull
-    to_date date str %m-%d-%Y - Ending date of the disclosure data pull
-    """
-    cookies = {
-        "BIGipServerPOOL_EDGE": "1427584378.20480.0000",
-        "JSESSIONID": "oAO1PNOZzGBoxIqtxy-32mVx.server-ep",
-    }
-
-    headers = {
-        "Origin": "https://edge.pse.com.ph",
-        "Accept-Encoding": "gzip, deflate",
-        "Accept-Language": "en-PH,en-US;q=0.9,en;q=0.8",
-        "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/74.0.3729.169 Safari/537.36",
-        "Content-Type": "application/x-www-form-urlencoded; charset=UTF-8",
-        "Accept": "*/*",
-        "Referer": "https://edge.pse.com.ph/announcements/form.do",
-        "X-Requested-With": "XMLHttpRequest",
-        "Connection": "keep-alive",
-    }
-
-    data = {
-        "companyId": "",
-        "keyword": symbol,
-        "tmplNm": "",
-        "fromDate": from_date,
-        "toDate": to_date,
-    }
-
-    response = requests.post(
-        "https://edge.pse.com.ph/announcements/search.ax",
-        headers=headers,
-        cookies=cookies,
-        data=data,
-    )
-    html = response.text
-    # Indicating the parser (e.g.  lxml) removes the bs warning
-    parsed_html = BeautifulSoup(html, "lxml")
-    table = parsed_html.find("table", {"class": "list"})
-    table_rows = table.find_all("tr")
-    lines = []
-    edge_nos = []
-    for tr in table_rows:
-        td = tr.find_all("td")
-        row = [tr.text for tr in td]
-        onclicks_raw = [
-            tr.a["onclick"]
-            for tr in td
-            if tr.a and "onclick" in tr.a.attrs.keys()
-        ]
-        onclicks = [s[s.find("('") + 2 : s.find("')")] for s in onclicks_raw]
-        lines.append(row)
-        if onclicks:
-            edge_nos.append(onclicks[0])
-
-    columns = [el.text for el in table.find_all("th")]
-
-    df = pd.DataFrame(lines, columns=columns)
-    # Filter to rows where not all columns are null
-    df = df[df.isna().mean(axis=1) < 1]
-    df["edge_no"] = edge_nos
-    df["url"] = (
-        "https://edge.pse.com.ph/openDiscViewer.do?edge_no=" + df.edge_no
-    )
-    df["Announce Date and Time"] = pd.to_datetime(df["Announce Date and Time"])
-    return df
-
-
-def get_disclosure_file_id(edge_no):
-    """
-    Returns file ID of a specified disclosure based on it edge number
-    ETA: 6.2 seconds per run
-    """
-    cookies = {
-        "BIGipServerPOOL_EDGE": "1427584378.20480.0000",
-        "JSESSIONID": "r2CYuOovD47c6FDnDoxHKW60.server-ep",
-    }
-
-    headers = {
-        "Connection": "keep-alive",
-        "Upgrade-Insecure-Requests": "1",
-        "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.88 Safari/537.36",
-        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
-        "Sec-Fetch-Site": "none",
-        "Sec-Fetch-Mode": "navigate",
-        "Accept-Encoding": "gzip, deflate, br",
-        "Accept-Language": "en-PH,en-US;q=0.9,en;q=0.8",
-    }
-
-    params = (("edge_no", edge_no),)
-
-    response = requests.get(
-        "https://edge.pse.com.ph/openDiscViewer.do",
-        headers=headers,
-        params=params,
-        cookies=cookies,
-    )
-    html = response.text
-    parsed_html = BeautifulSoup(html, "lxml")
-    s = parsed_html.iframe["src"]
-    file_id = s[s.find("file_id=") + 8 :]
-    return file_id
-
-
-def get_disclosure_parsed_html(disclosure_file_id):
-    """
-    Returns the bs parsed html for a disclosure given its file id
-    ETA: 6.55 seconds per run
-    """
-    cookies = {
-        "BIGipServerPOOL_EDGE": "1427584378.20480.0000",
-        "JSESSIONID": "r2CYuOovD47c6FDnDoxHKW60.server-ep",
-    }
-
-    headers = {
-        "Connection": "keep-alive",
-        "Upgrade-Insecure-Requests": "1",
-        "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.88 Safari/537.36",
-        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
-        "Sec-Fetch-Site": "same-origin",
-        "Sec-Fetch-Mode": "nested-navigate",
-        "Referer": "https://edge.pse.com.ph/openDiscViewer.do?edge_no=8a9a820ee365687cefdfc15ec263a54d",
-        "Accept-Encoding": "gzip, deflate, br",
-        "Accept-Language": "en-PH,en-US;q=0.9,en;q=0.8",
-    }
-
-    params = (("file_id", disclosure_file_id),)
-
-    response = requests.get(
-        "https://edge.pse.com.ph/downloadHtml.do",
-        headers=headers,
-        params=params,
-        cookies=cookies,
-    )
-    html = response.text
-
-    parsed_html = BeautifulSoup(html, "lxml")
-    return parsed_html
-
-
-def parse_stock_inventory(stock_inventory_str):
-    stock_inventory_lol = [
-        row.split("\n") for row in stock_inventory_str.split("\n\n\n\n")
-    ]
-    stock_inventory_df = pd.DataFrame(
-        stock_inventory_lol[1:], columns=stock_inventory_lol[0]
-    )
-    stock_inventory_df.iloc[:, 1] = (
-        stock_inventory_df.iloc[:, 1]
-        .apply(lambda x: x.replace(",", ""))
-        .astype(int)
-    )
-    return stock_inventory_df
-
-
-def get_company_summary(parsed_html):
-    """
-    Return the company summary (at the top) given the parsed html of the disclosure
-    """
-
-    keys = []
-    values = []
-    for dt, dd in zip(parsed_html.find_all("dt"), parsed_html.find_all("dd")):
-        # Take out first token (number followed by a period)
-        key = " ".join(dt.text.strip().split()[1:])
-        value = dd.text.strip()
-        if "Title of Each Class\n" in value:
-            stock_inventory_df = parse_stock_inventory(value)
-            keys += stock_inventory_df.iloc[:, 0].values.tolist()
-            values += stock_inventory_df.iloc[:, 1].values.tolist()
-        else:
-            keys.append(key)
-            values.append(value)
-
-    company_summary_df = pd.DataFrame()
-    company_summary_df["key"] = keys
-    company_summary_df["value"] = values
-    return company_summary_df
-
-
-def parse_table(table_el):
-    """
-    Returns a table as a dataframe from a table html element
-    """
-    table_dict = {"header": [], "value": []}
-    for tr in table_el.find_all("tr"):
-        th = None
-        td = None
-        if tr.find("th"):
-            th = tr.th.text
-        if tr.find("td"):
-            td = tr.td.text
-
-        table_dict["header"].append(th)
-        table_dict["value"].append(td)
-    return pd.DataFrame(table_dict)
-
-
-def get_tables(parsed_html):
-    """
-    Returns a list of tables as pd.DataFrame's from parsed HTML
-    """
-    table_els = parsed_html.find_all("table")
-    table_dfs = []
-    for table_el in table_els:
-        table_df = parse_table(table_el)
-        table_dfs.append(table_df)
-    return table_dfs
 
 
 def tweepy_api(consumer_key, consumer_secret, access_token, access_secret):
